Updated codecleandata.py
Will now adjust to the new file structures. No longer uses a set number of indexes but uses the variable names in the header of the csv file to find the correct data-columns. More user-friendly and easier to read. Plus it will update the user on its progress when script is running.